# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/recommendation": {
            "get": {
                "tags": [
                    "Recommendation"
                ],
                "summary": "Get list recommendation",
                "description": "Returns list of recommendation",
                "operationId": "getAllRecommendation",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Recommendation"
                ],
                "description": "Create new recommendation",
                "operationId": "dcefb46938ba284006320f804d14ddb4",
                "parameters": [
                    {
                        "name": "user_indicator",
                        "in": "query",
                        "description": "ID indicator",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_indicated",
                        "in": "query",
                        "description": "ID indicated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title of book",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success."
                    },
                    "422": {
                        "description": "Bad request."
                    }
                }
            }
        },
        "/api/recommendation/{recommendation}": {
            "get": {
                "tags": [
                    "Recommendation"
                ],
                "description": "Return recommendation by id - enabled",
                "operationId": "52baf9920f6aca5026fdd98b92a50b96",
                "parameters": [
                    {
                        "name": "recommendation",
                        "in": "path",
                        "description": "ID recommendation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success."
                    },
                    "422": {
                        "description": "User not found."
                    }
                }
            }
        },
        "/api/recommendation/{id}": {
            "delete": {
                "tags": [
                    "Recommendation"
                ],
                "summary": "Delete recommendation by id",
                "description": "Destroy recommendation",
                "operationId": "abd37bb3c9df5a12104e0d06797f7db6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID recommendation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success."
                    },
                    "401": {
                        "description": "permission denied."
                    },
                    "404": {
                        "description": "User not found."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of projects",
                "description": "Returns list of projects",
                "operationId": "getAllUser",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Create new user",
                "operationId": "089999749fbdec98922de2f74ba8d540",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "First name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Last name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success."
                    },
                    "422": {
                        "description": "Bad request."
                    }
                }
            }
        },
        "/api/user/{user}": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Return user by id - enabled",
                "operationId": "a0fa3a08f20fb611866a52d1bb2e43dd",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success."
                    },
                    "422": {
                        "description": "User not found."
                    }
                }
            }
        },
        "/api/user/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user by id",
                "description": "Destroy user",
                "operationId": "f8d7aea516ee464d0afee90f3a1a1722",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success."
                    },
                    "401": {
                        "description": "permission denied."
                    },
                    "404": {
                        "description": "User not found."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "security": [
        []
    ]
}